+++ ios/Framework/Core/WebContent/DisplayScreenlet/WebContentDisplayScreenlet.swift
@@ -44,6 +44,14 @@ public protocol WebContentDisplayScreenletDelegate: BaseScreenletDelegate {
 	@objc optional func screenlet(_ screenlet: WebContentDisplayScreenlet,
 			onWebContentError error: NSError)
 
+	/// Decides whether to allow or cancel a navigation.
+	///
+	///   - screenlet: Web content display screenlet instance.
+	///   - url: Clicked url.
+	/// - Returns: True if you're going to handle navigation, false otherwise.
+	@objc optional func screenlet(_ screenlet: WebContentDisplayScreenlet,
+								  onUrlClicked url: String) -> Bool
+
 }
 
 /// The Web Content Display Screenlet shows web content elements in your app, rendering the inner 
@@ -114,6 +122,13 @@ open class WebContentDisplayScreenlet: BaseScreenlet {
 
 				self.webContentDisplayViewModel?.customCssFile = self.customCssFile
 				self.webContentDisplayViewModel?.htmlContent = modifiedHtml ?? resultHtml
+				self.webContentDisplayViewModel?.onUrlClicked = { [weak self] url in
+					guard let strongSelf = self else {
+						return false
+					}
+					return strongSelf.webContentDisplayDelegate?.screenlet?(strongSelf, onUrlClicked: url) ?? false
+				}
+
 			}
 			else if let resultRecord = interactor.resultRecord {
 				self.webContentDisplayDelegate?.screenlet?(self,
@@ -139,5 +154,4 @@ open class WebContentDisplayScreenlet: BaseScreenlet {
 	open func loadWebContent() -> Bool {
 		return self.performDefaultAction()
 	}
-
 }