+++ ios/Framework/Themes/Default/WebContent/DisplayScreenlet/WebContentDisplayView_default.swift
@@ -15,7 +15,7 @@ import UIKit
 import WebKit
 
 @objc(WebContentDisplayView_default)
-open class WebContentDisplayView_default: BaseScreenletView, WebContentDisplayViewModel {
+open class WebContentDisplayView_default: BaseScreenletView, WebContentDisplayViewModel, WKNavigationDelegate {
 
 	// MARK: Outlets
 
@@ -23,7 +23,7 @@ open class WebContentDisplayView_default: BaseScreenletView, WebContentDisplayVi
 
 	open var injectedCss: String?
 
-	override open var progressMessages: [String:ProgressMessages] {
+	override open var progressMessages: [String: ProgressMessages] {
 		return [
 			BaseScreenlet.DefaultAction: [
 				.working: LocalizedString("default", key: "webcontentdisplay-loading-message", obj: self),
@@ -38,6 +38,8 @@ open class WebContentDisplayView_default: BaseScreenletView, WebContentDisplayVi
 		let config = WKWebViewConfiguration.noCacheConfiguration
 
 		webView = WKWebView(frame: self.frame, configuration: config)
+
+        webView?.navigationDelegate = self
 
 		webView?.injectCookies()
 		webView?.injectViewportMetaTag()
@@ -85,6 +87,28 @@ open class WebContentDisplayView_default: BaseScreenletView, WebContentDisplayVi
 			if let css = customCssFile {
 				injectedCss = webView?.loadCss(file: css)
 			}
+		}
+	}
+
+    open var onUrlClicked: ((String) -> Bool)?
+
+    // MARK: WKNavigationDelegate
+
+	public func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction,
+						decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+
+		if navigationAction.navigationType == WKNavigationType.linkActivated {
+			let webViewNotHandleLink = onUrlClicked?(webView.url?.absoluteString ?? "") ?? false
+
+			if webViewNotHandleLink {
+				decisionHandler(.cancel)
+			}
+			else {
+				decisionHandler(.allow)
+			}
+		}
+		else {
+			decisionHandler(.allow)
 		}
 	}
 }