+++ ios/Samples/Showcase-swift/LiferayScreens-Showcase-Swift/ViewControllers/WebContent/WebContentDisplayScreenletViewController.swift
@@ -50,6 +50,20 @@ class WebContentDisplayScreenletViewController: UIViewController, WebContentDisp
 		LiferayLogger.logDelegateMessage(args: error)
 	}
 	
+	func screenlet(_ screenlet: WebContentDisplayScreenlet,
+				   onUrlClicked url: String) -> Bool {
+		LiferayLogger.logDelegateMessage(args: url as AnyObject)
+		
+		//Return true to avoid opening the URL in WKWebView and decide what to do with it
+		//For example: open the URL in native browser
+		if let link = URL(string: url) {
+			UIApplication.shared.openURL(link)
+		}
+		return true
+		
+		//Return false to allow WKWebView navigate to the clicked link
+		//return false
+	}
 	
 	//MARK: UIViewController