+++ xamarin/Samples/AndorraTelecom-iOS/ICallMeBackDelegate.cs
@@ -1,0 +1,9 @@
+﻿namespace AndorraTelecomiOS
+{
+    public interface ICallMeBackDelegate
+    {
+        void ShowAlertLegalNotAccepted(CallMeBackView CallMeBackView, string Title, string Message);
+
+        void ShowLegalConditions(CallMeBackView CallMeBackView);
+    }
+}